Change: Move TeamPrivacy and TeamRole enums to base model module
